demo page review fixed
