+++ src/app/main/components/demo-page/demo-page.component.ts
@@ -1,21 +1,12 @@
-import { DropdownFieldType } from './../../../model/data-model';
+
+
+import { DropdownFieldType, Notification, Person } from './../../../model/data-model';
 import { ButtonType, ButtonThemeColor } from './../../../reusable/nv-button/nv-button.component';
 import { NvCustomComponent } from './../../../reusable/nv-custom-component/nv-custom-component.component';
-
 import { Component, OnInit } from '@angular/core';
 import { Button } from 'protractor';
 
-export interface Person {
-id: number;
-name: string;
-imageUrl: string;
-notification?: Notification;
-}
-export interface Notification {
-  number?: number | any;
-  content?: string;
-  hasAttachment?: boolean;
-}
+
 export enum NvComponent {
   // First: Add the name of your reusable component to this enum
   Search,
