+++ src/app/model/data-model.ts
@@ -3,3 +3,15 @@ export enum DropdownFieldType {
     Icon,
     Default,
 }
+
+export interface Person {
+    id: number;
+    name: string;
+    imageUrl: string;
+    notification?: Notification;
+}
+export interface Notification {
+    number?: number | any;
+    content?: string;
+    hasAttachment?: boolean;
+}
