+++ src/app/reusable/theme/base.scss
@@ -20,7 +20,7 @@ $color-light-blue-10: #C9E2F5;
 $color-light-blue-20: #F6FCFE;
 $color-black-10: #000000;
 $color-black: #1B1C1D;
-$color-black-10:0000001a;
+$color-black-10: #0000001a;
 $color-black-12: #00000066;
 $color-black-09 : #00000099;
 $color-black-14: #0000001A;
@@ -46,7 +46,6 @@ $color-light-gray-60: #333333DE;
 $color-light-gray-70: #9A9A9A;
 $color-light-gray-80: #8a8a8a;
 $color-light-gray-90: #999C9E;
-
 $color-light-gray-60: #333333DE;
 
 $color-white-10: #DEDEDF;
